moved code from fastestlap2.java to fastestlap.java and renamed class fastestlap
